flask_postgress_app/pyproject.toml
[project]
name = "flask-postgress-app"
version = "0.1.0"
description = "A simple Flask app with PostgreSQL"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "faker>=37.3.0",
    "flask==3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
